fixed error from homework

+++ nth_root_finder.h
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-int PowerNum(int number, int nth);
+long long int PowerNum(int number, int nth);
 long double PowerNum(double number, int nth);
 
 string FindNthRoot(int number, int nth, int precision);
